// packages/libs/types/store/centrifuge.ts
export enum CentrifugeActionTypes {
  Connect = 'CENTRIFUGE_CONNECT',
  Disconnect = 'CENTRIFUGE_DISCONNECT',
  Connected = 'CENTRIFUGE_CONNECTED',
  Disconnected = 'CENTRIFUGE_DISCONNECTED',
  Message = 'CENTRIFUGE_MESSAGE',
  SendMessage = 'CENTRIFUGE_SEND_MESSAGE',
  Subscribe = 'CENTRIFUGE_SUBSCRIBE',
  Unsubscribe = 'CENTRIFUGE_UNSUBSCRIBE',
  History = 'CENTRIFUGE_HISTORY',
  SetToken = 'CENTRIFUGE_SET_TOKEN',
}